Speed up multi-deletion
Take only a fine-grained lock on the store for deletion.  IIRC, the long
lock was to avoid conflict with Get which accesses the store in post_process(),
but this issue has been fixed and events only access the store in pre_process().
I believe it is now safe to only lock Delete during pre_process() as a result.

git-svn-id: http://svn.drobilla.net/lad/trunk/ingen@5771 a436a847-0d15-0410-975c-d299462d15a1