Always use sample accurate controls
This changes control port buffers to always be sequences of float, splitting
cycles internally so nodes with control ports act as if they support sample
accurate control input and output.  This allows things like having networks of
control ports manipulated by the user which are run into CV ports and having
sample accurate changes end up in the CV ports even though the values were
calculated by plugins with single float ports.

Further work is probably necessary to thin and smooth changes that come from a
user drag (perhaps there should be a mode for this?) to keep the amount of
cycle splitting reasonable, and support for plugins with fixed block length.