Set up serd to read directly from socket without a FILE intermediary
This was problematic because using fdopen() conflicts with poll().  If the FILE
ends up reading more than is actually processed by serd, then poll will not
fire because the socket file descriptor has been fully read, even though there
is pending input in the FILE object (which is buffered).

Avoid this by using a custom read function and calling recv() directly.  In
retrospect, this was stupid, but it seemed convenient to be able to use the
handy built-in support for reading from a FILE in serd.

Now, the client and server are using send() and recv() directly on the socket,
as things should be.  I am not sure if MSG_WAITALL is the best idea here, or if
it's really important at all, but it seems like a good idea.