Fix clashing blank nodes with MIDI bindings
This uses a different namespace for arc blank node IDs, which avoids clashes
with MIDI bindings which are written with IDs like "b1", "b2", and so on.  This
is a symptom of a deeper issue where blank node generation isn't coordinated
properly, but this fix avoids the problem for now and is nice anyway since it
makes the role of blank nodes clearer when debugging.