Avoid use of jack_frame_time()
This seems to be broken on the pipewire implementation of Jack, but we use
FrameTimer for this on other drivers anyway, so just use that for Jack as well
and avoid the issue entirely.

Conveniently also exercises the clock and timestamp DLL code, which as it turns
out, was itself broken.