Order dependencies to prioritize flags of lower level libraries
Specifically, to make the "system" include type for LV2 work and suppress C
warnings there.  Otherwise, the settings from other projects that depend on
it (like lilv) take priority.  This seems broken to me, and didn't used to
happen, so I'm not really sure what's changed, but this fixes it.